order service test fixed

+++ tests/Unit/API/Services/Coupon/CouponServiceTest.php
@@ -79,10 +79,10 @@ it('calculates discount for service fee order', function () {
     $discount = $this->couponService->calculateDiscount($this->order->coupon, $this->order);
 
     if ($this->coupon->type === 'fixed') {
-        $couponDiscount = $this->coupon->discount_value;
+        $couponDiscount = (float) $this->coupon->discount_value;
     } else {
         $serviceFee = $this->paymentPlan->price * array_sum(array_column($this->order->orderItems->toArray(), 'quantity'));
-        $couponDiscount = (($this->coupon->discount_value * $serviceFee) / 100);
+        $couponDiscount = (float) (($this->coupon->discount_value * $serviceFee) / 100);
     }
 
     expect($discount)->toBe($couponDiscount);
